fix(tui): remove StatusBar, move grants info to BudgetPanel, reverse history nav
