+++ loom/platform/cli/tui/__init__.py
@@ -21,7 +21,6 @@ from .components import (
     MessageBubble,
     ToolBlock,
     AgentState,
-    StatusBar,
     InputArea,
     ObservabilityPanel,
     ArtifactState,
@@ -57,7 +56,6 @@ __all__ = [
     "MessageBubble",
     "ToolBlock",
     "AgentState",
-    "StatusBar",
     "InputArea",
     "ObservabilityPanel",
     "ArtifactState",
