+++ loom/platform/cli/tui/components/__init__.py
@@ -7,7 +7,6 @@ from __future__ import annotations
 from .header import Header
 from .message_list import MessageList, MessageBubble
 from .tool_block import ToolBlock, AgentState
-from .status_bar import StatusBar
 from .input_area import InputArea
 from .observability_panel import ObservabilityPanel
 from .artifact_card import Artifact, ArtifactCard, ArtifactState
@@ -26,7 +25,6 @@ __all__ = [
     "MessageBubble",
     "ToolBlock",
     "AgentState",
-    "StatusBar",
     "InputArea",
     "ObservabilityPanel",
     "Artifact",
